Move `rand` to individual crates
