+++ bee-block/Cargo.toml
@@ -32,6 +32,7 @@ iterator-sorted = { version = "0.1.0", default-features = false }
 packable = { version = "0.5.0", default-features = false, features = [ "serde", "primitive-types" ] }
 prefix-hex = { version = "0.4.0", default-features = false, features = [ "primitive-types" ] }
 primitive-types = { version = "0.11.1", default-features = false, features = [ "serde" ] }
+rand = { version = "0.8.5", default-features = false, features = [ "min_const_gen" ], optional = true }
 serde = { version = "1.0.139", default-features = false, optional = true }
 serde-big-array = { version = "0.4.1", default-features = false, optional = true }
 serde_json = { version = "1.0.82", default-features = false, features = [ "alloc" ], optional = true }
@@ -46,3 +47,4 @@ num_cpus = { version = "1.13.1", default-features = false }
 dto = [ "serde", "dep:serde_json" ]
 serde = [ "dep:serde", "serde-big-array" ]
 std = [ ]
+rand = [ "dep:rand" ]
