+++ bee-block/src/lib.rs
@@ -33,6 +33,9 @@ pub mod output;
 pub mod parent;
 /// A module that provides types and syntactic validations of payloads.
 pub mod payload;
+/// A module that provides utilities for random generation of types.
+#[cfg(feature = "rand")]
+pub mod rand;
 /// A module that provides types and rules for semantic validation.
 pub mod semantic;
 /// A module that provides types and syntactic validations of signatures.
