+++ bee-block/src/rand/bool.rs
