+++ bee-block/src/rand/number.rs
