+++ bee-block/src/rand/payload.rs
@@ -1,23 +1,22 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::{
+use crate::{
     payload::{
         milestone::{MilestoneEssence, MilestoneOptions, MilestonePayload},
         tagged_data::TaggedDataPayload,
         treasury_transaction::TreasuryTransactionPayload,
         Payload,
     },
+    rand::{
+        bytes::{rand_bytes, rand_bytes_array},
+        input::rand_treasury_input,
+        milestone::{rand_merkle_root, rand_milestone_id, rand_milestone_index},
+        number::{rand_number, rand_number_range},
+        output::rand_treasury_output,
+        parents::rand_parents,
+    },
     signature::{Ed25519Signature, Signature},
-};
-
-use crate::rand::{
-    bytes::{rand_bytes, rand_bytes_array},
-    input::rand_treasury_input,
-    milestone::{rand_merkle_root, rand_milestone_id, rand_milestone_index},
-    number::{rand_number, rand_number_range},
-    output::rand_treasury_output,
-    parents::rand_parents,
 };
 
 /// Generates a random tagged data payload.
