+++ bee-block/src/rand/transaction.rs
@@ -1,9 +1,7 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::payload::transaction::TransactionId;
-
-use crate::rand::bytes::rand_bytes_array;
+use crate::{payload::transaction::TransactionId, rand::bytes::rand_bytes_array};
 
 /// Generates a random transaction id.
 pub fn rand_transaction_id() -> TransactionId {
