+++ bee-ledger/bee-ledger-types/src/rand/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+/// Module providing random output generation utilities.
+pub mod output;
+/// Module providing random output diff generation utilities.
+pub mod output_diff;
+/// Module providing random receipt generation utilities.
+pub mod receipt;
+/// Module providing random snapshot generation utilities.
+pub mod snapshot;
