+++ bee-ledger/bee-ledger-types/src/rand/receipt.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use bee_block::rand::{milestone::rand_milestone_index, milestone_option::rand_receipt_milestone_option};
+
+use crate::Receipt;
+
+/// Generates a random ledger receipt.
+pub fn rand_ledger_receipt() -> Receipt {
+    Receipt::new(rand_receipt_milestone_option(), rand_milestone_index())
+}
