+++ bee-storage/bee-storage-test/Cargo.toml
@@ -17,10 +17,10 @@ all-features = true
 rustdoc-args = [ "--cfg", "doc_cfg" ]
 
 [dependencies]
-bee-block = { version = "1.0.0-beta.2", path = "../../bee-block", default-features = false }
+bee-block = { version = "1.0.0-beta.2", path = "../../bee-block", default-features = false, features = [ "rand" ] }
 bee-ledger = { version = "0.7.0", path = "../../bee-ledger/bee-ledger", default-features = false }
+bee-ledger-types = { version = "1.0.0-beta.1", path = "../../bee-ledger/bee-ledger-types", default-features = false, features = [ "rand" ] }
 bee-storage = { version = "0.12.0", path = "../bee-storage", default-features = false }
-bee-tangle = { version = "0.3.0", path = "../../bee-tangle", default-features = false }
-bee-test = { version = "0.1.0", path = "../../bee-test", default-features = false }
+bee-tangle = { version = "0.3.0", path = "../../bee-tangle", default-features = false, features = [ "rand" ] }
 
 packable = { version = "0.5.0", default-features = false, features = [ "serde" ] }
