+++ bee-storage/bee-storage-test/src/milestone_index_to_milestone_metadata.rs
@@ -1,13 +1,12 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::payload::milestone::MilestoneIndex;
+use bee_block::{payload::milestone::MilestoneIndex, rand::milestone::rand_milestone_index};
 use bee_storage::{
     access::{AsIterator, Batch, BatchBuilder, Delete, Exist, Fetch, Insert, MultiFetch, Truncate},
     backend,
 };
-use bee_tangle::milestone_metadata::MilestoneMetadata;
-use bee_test::rand::milestone::{rand_milestone_index, rand_milestone_metadata};
+use bee_tangle::{milestone_metadata::MilestoneMetadata, rand::milestone::rand_milestone_metadata};
 
 pub trait StorageBackend:
     backend::StorageBackend
