+++ bee-storage/bee-storage-test/src/solid_entry_point_to_milestone_index.rs
@@ -1,13 +1,12 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::payload::milestone::MilestoneIndex;
+use bee_block::{payload::milestone::MilestoneIndex, rand::milestone::rand_milestone_index};
 use bee_storage::{
     access::{AsIterator, Batch, BatchBuilder, Delete, Exist, Fetch, Insert, MultiFetch, Truncate},
     backend,
 };
-use bee_tangle::solid_entry_point::SolidEntryPoint;
-use bee_test::rand::{milestone::rand_milestone_index, solid_entry_point::rand_solid_entry_point};
+use bee_tangle::{rand::solid_entry_point::rand_solid_entry_point, solid_entry_point::SolidEntryPoint};
 
 pub trait StorageBackend:
     backend::StorageBackend
