+++ bee-tangle/Cargo.toml
@@ -17,7 +17,8 @@ all-features = true
 rustdoc-args = [ "--cfg", "doc_cfg" ]
 
 [dependencies]
-bee-block = { version = "1.0.0-beta.2", path = "../bee-block", default-features = false, features = [ "serde" ] }
+bee-block = { version = "1.0.0-beta.2", path = "../bee-block", default-features = false, features = [ "serde", "rand" ] }
+bee-ledger-types = { version = "1.0.0-beta.1", path = "../bee-ledger/bee-ledger-types", default-features = false, features = [ "rand" ], optional = true }
 bee-runtime = { version = "0.1.1-alpha", path = "../bee-runtime", default-features = false }
 bee-storage = { version = "0.12.0", path = "../bee-storage/bee-storage", default-features = false }
 
@@ -41,6 +42,9 @@ bee-test = { path = "../bee-test", default-features = false }
 criterion = { version = "0.3.5", default-features = false, features = [ "async_tokio" ] }
 rand = { version = "0.8.5", default-features = false }
 tokio = { version = "1.19.2", default-features = false, features = [ "rt", "rt-multi-thread", "sync", "time" ] }
+
+[features]
+rand = [ "dep:bee-ledger-types" ]
 
 [[bench]]
 name = "tangle_bench"
