+++ bee-tangle/benches/tangle_bench.rs
@@ -1,11 +1,14 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::{semantic::ConflictReason, Block, BlockId};
+use bee_block::{
+    rand::{block::rand_block, block_metadata::rand_block_metadata, number::rand_number},
+    semantic::ConflictReason,
+    Block, BlockId,
+};
 use bee_runtime::resource::ResourceHandle;
 use bee_storage_null::Storage as NullStorage;
 use bee_tangle::{block_metadata::BlockMetadata, config::TangleConfig, Tangle};
-use bee_test::rand::{block::rand_block, block_metadata::rand_block_metadata, number::rand_number};
 use criterion::*;
 use rand::seq::SliceRandom;
