+++ bee-tangle/src/rand/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+/// Module providing random block metadata generation utilities.
+pub mod block_metadata;
+/// Module providing random milestone generation utilities.
+pub mod milestone;
+/// Module providing random solid entry point generation utilities.
+pub mod solid_entry_point;
+/// Module providing random unreferenced block generation utilities.
+pub mod unreferenced_block;
