+++ bee-tangle/src/rand/solid_entry_point.rs
@@ -1,9 +1,9 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_tangle::solid_entry_point::SolidEntryPoint;
+use bee_block::rand::block::rand_block_id;
 
-use crate::rand::block::rand_block_id;
+use crate::solid_entry_point::SolidEntryPoint;
 
 /// Generates a random solid entry point.
 pub fn rand_solid_entry_point() -> SolidEntryPoint {
